requirements.txt
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.3
djangorestframework-gis==0.12
psycopg2==2.7.3.2
coreapi=2.3.3


